auth/requirements.txt
yubico-client
pycryptodomex